refactor: settlement module 추출
